refactor: [UNSTABLE] refactor pointer usage
Updates the compiler.zig, data.zig files and tests to use allocators
for strings to respect lifetimes better. this fixes a class of segfaults
I was getting when integrating with the ast struct.

More updates to come into function.zig, ast.zig and tests so this commit
leaves the repo in a unstable state, but the tests and build pass